Trait elface::MetaData

pub trait MetaData<T: AsRef<[u8]> + Sized>: Sized {
    fn as_slice<'a>(&'a self) -> &'a [u8] { ... }

    fn as_mut_slice<'a>(&'a mut self) -> &'a mut [u8] { ... }

    fn read_from_slice(&mut self, src: &[u8]) { ... }

    fn to_be_bytes(self) -> T { ... }

    fn to_le_bytes(self) -> T { ... }

    fn to_ne_bytes(self) -> T { ... }

    unsafe fn from_be_bytes(_: T) -> Self { ... }

    unsafe fn from_le_bytes(_: T) -> Self { ... }

    fn from_ne_bytes(_: T) -> Self { ... }

    fn from_be(_: Self) -> Self { ... }

    fn from_le(_: Self) -> Self { ... }

    fn to_be(self) -> Self { ... }

    fn to_le(self) -> Self { ... }

    fn len() -> usize { ... }
}

Provided methods

fn as_slice<'a>(&'a self) -> &'a [u8]

fn as_mut_slice<'a>(&'a mut self) -> &'a mut [u8]

fn read_from_slice(&mut self, src: &[u8])

fn to_be_bytes(self) -> T

将 self 转换为大端的字节数组， 以及类似的字节端转换都会消耗源值， 这是因为转换前和转换后是两个完全不同的数据表示

fn to_le_bytes(self) -> T

fn to_ne_bytes(self) -> T

unsafe fn from_be_bytes(_: T) -> Self

Read from bytes as if it's big endianess. It's marked as unsafe since we can not make sure the source's value is in valid range. You should use it as the source from to_be_bytes.

unsafe fn from_le_bytes(_: T) -> Self

Read from bytes as if it's little endianess. It's marked as unsafe since we can not make sure the source's value is in valid range. You should use it as the source from to_le_bytes.

fn from_ne_bytes(_: T) -> Self

fn from_be(_: Self) -> Self

fn from_le(_: Self) -> Self

fn to_be(self) -> Self

fn to_le(self) -> Self

fn len() -> usize

Loading content...